backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
ansicon==1.89.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
awpy==2.0.0b4
billiard==4.2.1
blessed==1.20.0
celery==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
demoparser2==0.34.0
Django==5.1.4
django-apscheduler==0.7.0
django-bootstrap5==25.1
django-cors-headers==4.6.0
django-debug-toolbar==5.1.0
django-picklefield==3.3
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
eventlet==0.39.1
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
jinxed==1.3.0
kiwisolver==1.4.8
kombu==5.5.2
loguru==0.7.3
matplotlib==3.10.0
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
polars==1.19.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
pyarrow==18.1.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
usd-core==24.11
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
yarl==1.18.3